% Generated by roxygen2: do not edit by hand
\docType{package}
\name{dsPUcopulaClient-package}
\alias{dsPUcopulaClient}
\alias{dsPUcopulaClient-package}
\title{Client-side utilities for the dsPUcopula DataSHIELD package}
\description{
Client-facing helpers for orchestrating the remote procedures exposed by the
\pkg{dsPUcopula} server package. The functions simplify typical workflows for
partition-of-unity copula modelling in a DataSHIELD environment.
}
\author{
\strong{Maintainer}: Andreas M\enc{\"a}{a}ndle \email{andreas.maendle@leibniz-bips.de}
}
\keyword{internal}
